+++ game/g_spawn.c
@@ -326,7 +326,7 @@ ED_NewString
 char *ED_NewString (char *string)
 {
 	char	*newb, *new_p;
-	int		i,l;
+	size_t	i,l;
 	
 	l = strlen(string) + 1;
